requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.2.0
matplotlib==3.7.1
seaborn==0.12.2
tqdm==4.65.0
python-dotenv==1.0.0
scipy==1.10.1
requests==2.28.2
torch==2.1.2
tensorflow==2.15.0
xgboost==2.0.3
optuna==3.5.0
